requirements.txt
pydantic>=2.6.1
python-dotenv>=1.0.0
typing-extensions>=4.9.0
requests>=2.31.0
pytest>=7.4.0 